Display tvl on the Overview page
