+++ src/pages/Overview/Network/index.tsx
@@ -7,9 +7,7 @@ import UpgradeBanner from "./UpgradeBanner"
 import WalletBalances from "./WalletBalances"
 import Nodes from "./Nodes"
 
-const Network: FC = () => {
-  const totalValueLocked = 30838938
-
+const Network: FC<{ totalValueLocked: string }> = ({ totalValueLocked }) => {
   return (
     <VStack spacing={4} mt={4}>
       <UpgradeBanner />
